+++ TP_MAPREDUCE/src/main/java/org/example/TotalPriceReducer.java
@@ -1,5 +1,6 @@
 package org.example;
 
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
@@ -7,14 +8,14 @@ import org.apache.hadoop.mapreduce.Reducer;
 import java.io.IOException;
 import java.util.Iterator;
 
-public class TotalPriceReducer extends Reducer<Text, IntWritable,Text,IntWritable> {
+public class TotalPriceReducer extends Reducer<Text, DoubleWritable,Text,DoubleWritable> {
     @Override
-    protected void reduce(Text key, Iterable<IntWritable> values, Reducer<Text, IntWritable, Text, IntWritable>.Context context) throws IOException, InterruptedException {
-       Iterator<IntWritable> it=values.iterator();
-       int somme=0;
+    protected void reduce(Text key, Iterable<DoubleWritable> values, Reducer<Text, DoubleWritable, Text, DoubleWritable>.Context context) throws IOException, InterruptedException {
+       Iterator<DoubleWritable> it=values.iterator();
+       double somme=0;
        while (it.hasNext()){
            somme+=it.next().get();
        }
-       context.write(key,new IntWritable(somme));
+       context.write(key, new DoubleWritable(somme));
     }
 }
